requirements.txt
tensorflow
numpy
opencv-python
PyYAML
scikit-learn
pandas
matplotlib